daily-light-backend/requirements.txt
Flask>=2.0
openai>=1.0 # For OpenAI API access
python-dotenv>=0.19 # To load .env files for local development 